migrate all the determinant-related tests to test_determinant.py
